requirements.txt
python-dotenv
semantic-kernel
azure-ai-inference